STORE-945 v2.1 Maintenance

+++ server/openstorefront/openstorefront-web/src/main/java/edu/usu/sdl/openstorefront/web/test/alert/AlertUserDataTest.java
@@ -51,19 +51,25 @@ public class AlertUserDataTest extends BaseTestCase
 	private Alert alertUserDataTag = null;
 	private Alert alertUserDataRQ = null;
 	private Alert alertUserDataContact = null;
-	private List<Alert> alerts = null;
-	private ExperienceTimeType experience = null;
-	String messageId = null;
-	String contactId = null;
+	private List<Alert> existingAlerts = null;
+	private ExperienceTimeType experienceTimeType = null;
+	private String messageId = null;
+	private String contactId = null;
+
+	@Override
+	protected void initializeTest()
+	{
+		// Get all alerts and set to inactive
+		existingAlerts = getActiveAlerts();
+		if (!existingAlerts.isEmpty()) {
+			inactivateAlerts(existingAlerts);
+		}
+	}
 
 	@Override
 	protected void runInternalTest()
 	{
-		// Get all alerts and set to inactive
-		alerts = getActiveAlerts();
-		if (!alerts.isEmpty()) {
-			inactivateAlerts(alerts);
-		}
+		ComponentAll componentAll = getTestComponent();
 
 		results.append("Test 1 - One User Data Option Selected<br><br>");
 		alertUserDataTag = new Alert();
@@ -84,13 +90,12 @@ public class AlertUserDataTest extends BaseTestCase
 		results.append("Alert saved<br><br>");
 
 		results.append("Checking alert user data tag message...<br>");
-		ComponentAll componentAllTag = getTestComponent();
-		Component componentAlertTag = componentAllTag.getComponent();
+		Component componentAlertTag = componentAll.getComponent();
 		ComponentTag compTag = new ComponentTag();
 		compTag.setText("New Test Tag");
 		compTag.setComponentId(componentAlertTag.getComponentId());
 		compTag.setAdminModified(false);
-		componentAllTag.getTags().add(compTag);
+		componentAll.getTags().add(compTag);
 		service.getComponentService().saveComponentTag(compTag);
 
 		UserMessage userMessage = new UserMessage();
@@ -127,17 +132,16 @@ public class AlertUserDataTest extends BaseTestCase
 		alertUserDataRQ = service.getAlertService().saveAlert(alertUserDataRQ);
 
 		results.append("Checking alert user data question message...<br>");
-		ComponentAll componentAllRQ = getTestComponent();
-		Component compAlertQuestion = componentAllRQ.getComponent();
+
 		QuestionAll question = new QuestionAll();
 		ComponentQuestion componentQuestion = new ComponentQuestion();
-		componentQuestion.setComponentId(compAlertQuestion.getComponentId());
+		componentQuestion.setComponentId(componentAll.getComponent().getComponentId());
 		componentQuestion.setUserTypeCode(END_USER);
 		componentQuestion.setQuestion("What is the largest basin on the moon?");
 		componentQuestion.setOrganization(TEST_ORGANIZATION);
-		compAlertQuestion.setAdminModified(false);
+		componentQuestion.setAdminModified(false);
 		question.setQuestion(componentQuestion);
-		componentAllRQ.getQuestions().add(question);
+		componentAll.getQuestions().add(question);
 		service.getComponentService().saveComponentQuestion(componentQuestion);
 
 		userMessage = new UserMessage();
@@ -164,7 +168,7 @@ public class AlertUserDataTest extends BaseTestCase
 		ReviewAll review = new ReviewAll();
 		ComponentReview componentReview = new ComponentReview();
 		componentReview.setUserTypeCode(END_USER);
-		componentReview.setComponentId(compAlertQuestion.getComponentId());
+		componentReview.setComponentId(componentAll.getComponent().getComponentId());
 		componentReview.setComment("User Data Review Test Comment");
 		componentReview.setRating(4);
 		componentReview.setTitle("User Data Review Test Title");
@@ -173,12 +177,14 @@ public class AlertUserDataTest extends BaseTestCase
 		componentReview.setAdminModified(false);
 		componentReview.setUserTimeCode("TIME_TEST");
 		componentReview.setLastUsed(TimeUtil.currentDate());
-		experience = new ExperienceTimeType();
-		experience.setCode("TIME_TEST");
-		experience.setDescription("The Time Experience Test");
-		service.getLookupService().saveLookupValue(experience);
+
+		experienceTimeType = new ExperienceTimeType();
+		experienceTimeType.setCode("TIME_TEST");
+		experienceTimeType.setDescription("The Time Experience Test");
+		service.getLookupService().saveLookupValue(experienceTimeType);
+
 		review.setComponentReview(componentReview);
-		componentAllRQ.getReviews().add(review);
+		componentAll.getReviews().add(review);
 		service.getComponentService().saveComponentReview(componentReview);
 
 		userMessage = new UserMessage();
@@ -214,25 +220,24 @@ public class AlertUserDataTest extends BaseTestCase
 		alertUserDataContact = service.getAlertService().saveAlert(alertUserDataContact);
 
 		results.append("Checking alert contact message...<br>");
-		ComponentAll componentAllContact = getTestComponent();
-		Component compContact = componentAllContact.getComponent();
+
 		ComponentContact compAlertContact = new ComponentContact();
 		Contact contact = new Contact();
 		compAlertContact.setContactId(contact.getContactId());
-		compAlertContact.setComponentId(compContact.getComponentId());
+		compAlertContact.setComponentId(componentAll.getComponent().getComponentId());
 		compAlertContact.setContactType(SUBMITTER);
 		compAlertContact.setFirstName("UserDataFirstName1");
 		compAlertContact.setLastName("UserDataLastName1");
 		compAlertContact.setEmail("myUserDataTest@alertuserdata.com");
 		compAlertContact.setOrganization(TEST_ORGANIZATION);
 		compAlertContact.setAdminModified(false);
-		componentAllContact.getContacts().add(compAlertContact);
+		componentAll.getContacts().add(compAlertContact);
 		service.getComponentService().saveComponentContact(compAlertContact);
 
 		ComponentContact updatedCompContact = new ComponentContact();
 		updatedCompContact.setFirstName("UserDataFirstName1");
 		updatedCompContact.setEmail("myUserDataTest@alertuserdata.com");
-		updatedCompContact.setComponentId(compContact.getComponentId());
+		updatedCompContact.setComponentId(componentAll.getComponent().getComponentId());
 		updatedCompContact = updatedCompContact.find();
 		contactId = updatedCompContact.getContactId();
 		updatedCompContact.setFirstName("Updated UserData Contact FirstName");
@@ -252,7 +257,6 @@ public class AlertUserDataTest extends BaseTestCase
 		if (!alertIdsEqual) {
 			results.append("Test Passed - User data message not found<br><br>");
 		} else {
-			service.getUserService().processAllUserMessages(true);
 			service.getUserService().removeUserMessage(messageId);
 			failureReason.append("Test Failed - User data message found<br><br>");
 		}
@@ -274,15 +278,17 @@ public class AlertUserDataTest extends BaseTestCase
 		if (alertUserDataContact != null) {
 			service.getAlertService().deleteAlert(alertUserDataContact.getAlertId());
 		}
-		if (experience != null) {
-			service.getLookupService().removeValue(ExperienceTimeType.class, experience.getCode());
+		if (experienceTimeType != null) {
+			service.getLookupService().removeValue(ExperienceTimeType.class, experienceTimeType.getCode());
 		}
 		if (contactId != null) {
 			service.getContactService().deleteContact(contactId);
 		}
-		service.getUserService().removeUserMessage(messageId);
-		if (!alerts.isEmpty()) {
-			activateAlerts(alerts);
+		if (messageId != null) {
+			service.getUserService().removeUserMessage(messageId);
+		}
+		if (!existingAlerts.isEmpty()) {
+			activateAlerts(existingAlerts);
 		}
 	}
